+++ packages/quick-question/services/RepositoryManager.ts
@@ -1,6 +1,7 @@
 import fs from "fs";
 import path from "path";
 
+import { globSync } from "glob";
 import { buildIndex } from "quick-question-indexer";
 
 export interface Metadata {
@@ -10,32 +11,30 @@ export interface Metadata {
 
 export type IndexingStatus = "init" | "success" | "pending" | "failed";
 
-class RepositoryManager {
-  indexingJob?: Promise<void>;
-  indexingStatus: IndexingStatus = "init";
+export interface ProjectInfo {
   metadata: Metadata;
+  indexingStatus: IndexingStatus;
+}
 
-  constructor() {
-    const metadataFile = path.join(
-      process.cwd(),
-      process.env.REPO_DIR!,
-      "metadata.json"
-    );
+class Project {
+  readonly projectDir: string;
+  readonly metadata: Metadata;
+  indexingStatus: IndexingStatus = "init";
+  indexingJob?: Promise<void>;
+
+  constructor(metadataFile) {
+    this.projectDir = path.dirname(metadataFile);
     this.metadata = JSON.parse(fs.readFileSync(metadataFile, "utf-8"));
   }
 
-  getIndexingStatus(): IndexingStatus {
-    if (
-      fs.existsSync(
-        path.join(process.cwd(), process.env.REPO_DIR!, "index/docstore.json")
-      )
-    ) {
+  private fetchIndexingStatus(dryrun: boolean): IndexingStatus {
+    if (fs.existsSync(path.join(this.projectDir, "index/docstore.json"))) {
       return "success";
     }
 
-    if (!this.indexingJob) {
+    if (!this.indexingJob && !dryrun) {
       this.indexingStatus = "pending";
-      this.indexingJob = createIndexingJob().then((status) => {
+      this.indexingJob = createIndexingJob(this.projectDir).then((status) => {
         this.indexingStatus = status;
       });
     }
@@ -44,16 +43,53 @@ class RepositoryManager {
   }
 }
 
-async function createIndexingJob(): Promise<IndexingStatus> {
+async function createIndexingJob(projectDir): Promise<IndexingStatus> {
   try {
     await buildIndex({
-      input: path.join(process.cwd(), process.env.REPO_DIR!, "metadata.json"),
+      input: path.join(projectDir, "metadata.json"),
       dryrun: false,
     });
     return "success";
   } catch (err) {
     console.error("Failed indexing", err);
     return "failed";
+  }
+}
+
+class RepositoryManager {
+  private readonly projects: Project[];
+
+  constructor() {
+    this.projects = globSync(
+      path.join(process.env.REPO_DIR!, "*", "metadata.json")
+    ).map((x) => new Project(x));
+  }
+
+  collectProjectInfos(): ProjectInfo[] {
+    return this.projects.map((x) => ({
+      metadata: x.metadata,
+      indexingStatus: x.fetchIndexingStatus(true),
+    }));
+  }
+
+  fetchProjectInfo(name: string): ProjectInfo {
+    for (const x of this.projects) {
+      if (x.metadata.name === name)
+        return {
+          metadata: x.metadata,
+          indexingStatus: x.fetchIndexingStatus(false),
+        };
+    }
+
+    throw new Error("Invalid project name " + name);
+  }
+
+  getProject(name: string): Project {
+    for (const x of this.projects) {
+      if (x.metadata.name === name) return x;
+    }
+
+    throw new Error("Invalid project name " + name);
   }
 }
